adds join mode
The join mode treats every sequences from a file as part of one genome. It does
so by appending the sequences to each other with a separator character in between.
This is useful if contigs of one organism are kept in one file and need to be
compared to another organism.

The join mode is activated via --join or -j and expects at least one filename
to be passed via command line arguments.