refactors all the code to plain C
As the RMQs were the only dependency using C++, and we got rid of those,
I refactored the code to plain C11. Thus we no longer need a C++ compiler,
except for the unit tests, or psufsort. This might make andi easier to port
to old machines without decent C++ compilers. Additionally, this now allows
us to use the cool C features, missing in C++.