release v0.9
This is version 0.9 of the `andi` tool for efficient estimation of evolutionary distances. It is faster and uses less memory than the last version. It also features a new suffix array construction algorithm: psufsort. That one can be activated via:

    ./configure --without-libdivsufsort

Also, a lot of the code has been refactored to plain C. This reduces the dependencies and makes the code easier to maintain.