Adds Manual
This commit adds the user manual for andi. It is written in LaTeX
and thus compiles to PDF. Some images, which are not generated via
LaTeX them selves, are given as .pdf. Further more .tex and .data
files are given for advanced plots.

Even though, one should not include generated files into source
control, I include andi-manual.pdf. This way, I can link to it online.
Also, a the user might not have LaTeX or the neccessary packages, to
build the manual. Thus, the final result, but not intermediate steps,
are committed.