adds distance cutoff
For closely related clades its good enough to just compare taxa within and
pick one representing taxon fo comparisons with outgroups. This reduces the
number of comparisons for distant genomes, and thus speeds up the computation.
A global variable called CUTOFF is used as a threshold to determine, when
a distance should be estimated and not computed.