Fixes integer underflow
This kind of reverts 3bb40b which in turn, reverted 596771c. I was right
afterall, that I have to ensure that matches are increasing. But that is
guaranteed in the query, but not in the subject. So that is the place where
I have to check for a possible underflow.