add file-of-filenames argument
To circumvent the limits imposed by the shell on the length of a command,
andi now can read a list of files to analyse from a file itself. This is
achieved via the new `--file-of-filenames` parameter with the argument
being the fof path.

With this new feature,

    andi snippets.fasta

is now equivalent to

    echo "snippets.fasta" > fof.txt
    andi --file-of-filenames fof.txt

Furthermore, andi can also read filenames via stdin, leading to the
following pattern

    ls -d -1 /path/to/*.fasta | andi --file-of-filenames - --join

Requested by @schultzm.
See also issue #7.