implement lucky anchors
Usually anchors are required to be unique and longer than the
threshold. However, anchors always come in pairs. Thus if the left
anchor is already unique we might drop this criterion for the right
anchor. This allows us to skip the matching step via the suffix array
and directly compare nucleotides instead. If lucky anchor is too short,
fall back to the normal strategy.

As we skip the expensive matching step, this method finds anchors much
faster, given they are only separated by a single SNP. For Eco29 this is
the case in roughly 60% of the anchors. In total the runtime decreases
by 10% to 15%.